moved MaximumSlippage to own file
1. minor css changes
